Reorganize test directories
More clearly separate unit and integration tests
Allow splitting into multiple files
Cleaner
