+++ tests/integration/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from .api_test import *
